Split processors into separate classes.  Performing bean injection from blueprint context file
