+++ services-camel/pom.xml
@@ -1,45 +1,58 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
-  <modelVersion>4.0.0</modelVersion>
+	<modelVersion>4.0.0</modelVersion>
 
-  <parent>
-    <artifactId>cdr</artifactId>
-    <groupId>edu.unc.lib.cdr</groupId>
-    <version>4.0-SNAPSHOT</version>
-  </parent>
+	<parent>
+		<artifactId>cdr</artifactId>
+		<groupId>edu.unc.lib.cdr</groupId>
+		<version>4.0-SNAPSHOT</version>
+	</parent>
 
-  <artifactId>services-camel</artifactId>
-  <version>1.0</version>
-  <packaging>jar</packaging>
+	<artifactId>services-camel</artifactId>
+	<version>1.0</version>
+	<packaging>jar</packaging>
 
-  <name>thumbnail-enhancements</name>
+	<name>thumbnail-enhancements</name>
 
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-    <fcrepo-camel.version>4.4.4</fcrepo-camel.version>
-    <camel.version>2.17.3</camel.version>
-  </properties>
+	<properties>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
+		<fcrepo-camel.version>4.4.4</fcrepo-camel.version>
+		<camel.version>2.17.3</camel.version>
+		<activemq.version>5.14.1</activemq.version>
+	</properties>
 
-  <dependencies>
+	<dependencies>
 
-    <!-- Camel -->
-    <dependency>
-      <groupId>org.apache.camel</groupId>
-      <artifactId>camel-core</artifactId>
-      <version>${camel.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.camel</groupId>
-      <artifactId>camel-blueprint</artifactId>
-      <version>${camel.version}</version>
-    </dependency>
+	<!-- Camel -->
+	<dependency>
+		<groupId>org.apache.camel</groupId>
+		<artifactId>camel-core</artifactId>
+		<version>${camel.version}</version>
+	</dependency>
+	<dependency>
+		<groupId>org.apache.camel</groupId>
+		<artifactId>camel-blueprint</artifactId>
+		<version>${camel.version}</version>
+	</dependency>
+	<dependency>
+		<groupId>org.apache.camel</groupId>
+		<artifactId>camel-exec</artifactId>
+		<version>${camel.version}</version>
+	</dependency>
 
-    <dependency>
-			<groupId>org.fcrepo.camel</groupId>
-			<artifactId>fcrepo-camel</artifactId>
-			<version>${fcrepo-camel.version}</version>
+	<dependency>
+		<groupId>org.fcrepo.camel</groupId>
+		<artifactId>fcrepo-camel</artifactId>
+		<version>${fcrepo-camel.version}</version>
+	</dependency>
+
+	<dependency>
+		<groupId>org.apache.activemq</groupId>
+		<artifactId>activemq-camel</artifactId>
+		<version>${activemq.version}</version>
+		<scope>test</scope>
 	</dependency>
 	
 	<dependency>
@@ -50,12 +63,6 @@
 	</dependency>
 	
 	<dependency>
-		  <groupId>com.google.guava</groupId>
-		  <artifactId>guava</artifactId>
-		  <version>19.0</version>
-	</dependency>
-	
-	<dependency>
 		<groupId>edu.unc.lib.cdr</groupId>
 		<artifactId>metadata</artifactId>
 	</dependency>
@@ -63,39 +70,60 @@
 	<dependency>
 		<groupId>edu.unc.lib.cdr</groupId>
 		<artifactId>fcrepo-clients</artifactId>
+		<exclusions>
+			<exclusion>
+				<artifactId>org.osgi.core</artifactId>
+				<groupId>org.osgi</groupId>
+			</exclusion>
+		</exclusions>
 	</dependency>
 
-    <!-- Testing -->
-    <dependency>
-      <groupId>org.apache.camel</groupId>
-      <artifactId>camel-test-blueprint</artifactId>
-      <version>2.17.3</version>
-      <scope>test</scope>
-    </dependency>
-    
-    <dependency>
+	<!-- Testing -->
+	<dependency>
+		<groupId>org.apache.camel</groupId>
+		<artifactId>camel-test-blueprint</artifactId>
+		<version>${camel.version}</version>
+		<scope>test</scope>
+	</dependency>
+	<dependency>
+		<groupId>org.apache.camel</groupId>
+		<artifactId>camel-test-spring</artifactId>
+		<version>${camel.version}</version>
+		<scope>test</scope>
+	</dependency>
+
+	<dependency>
 		<groupId>edu.unc.lib.cdr</groupId>
 		<artifactId>metadata</artifactId>
 		<type>test-jar</type>
 		<scope>test</scope>
 	</dependency>
-    
-    <!-- logging -->
- <!--    <dependency>
-      <groupId>org.apache.logging.log4j</groupId>
-      <artifactId>log4j-api</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.logging.log4j</groupId>
-      <artifactId>log4j-core</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.logging.log4j</groupId>
-      <artifactId>log4j-slf4j-impl</artifactId>
-      <scope>test</scope>
-    </dependency>--> 
+	
+	<dependency>
+		<groupId>org.mockito</groupId>
+		<artifactId>mockito-all</artifactId>
+		<scope>test</scope>
+	</dependency>
+
+	<!-- logging -->
+	<dependency>
+		<groupId>org.apache.logging.log4j</groupId>
+		<artifactId>log4j-api</artifactId>
+		<version>2.6.2</version>
+		<scope>test</scope>
+	</dependency>
+	<dependency>
+		<groupId>org.apache.logging.log4j</groupId>
+		<artifactId>log4j-core</artifactId>
+		<version>2.6.2</version>
+		<scope>test</scope>
+	</dependency>
+	<dependency>
+		<groupId>org.apache.logging.log4j</groupId>
+		<artifactId>log4j-slf4j-impl</artifactId>
+		<version>2.6.2</version>
+		<scope>test</scope>
+	</dependency>
 
   </dependencies>
 
@@ -112,47 +140,11 @@
           <target>1.8</target>
         </configuration>
       </plugin>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-resources-plugin</artifactId>
-        <version>3.0.1</version>
-        <configuration>
-          <encoding>UTF-8</encoding>
-        </configuration>
-      </plugin>
 
       <!-- to generate the MANIFEST.MF of the bundle -->
       <plugin>
         <groupId>org.apache.felix</groupId>
         <artifactId>maven-bundle-plugin</artifactId>
-        <version>3.2.0</version>
-        <executions>
-          <execution>
-            <id>bundle-manifest</id>
-            <phase>prepare-package</phase>
-            <goals>
-              <goal>manifest</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-      <!-- to include MANIFEST.MF in the bundle -->
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-jar-plugin</artifactId>
-        <version>2.6</version>
-        <configuration>
-          <archive>
-            <manifestFile>${project.build.outputDirectory}/META-INF/MANIFEST.MF</manifestFile>
-          </archive>
-        </configuration>
-      </plugin>
-
-      <!-- to run the example using mvn camel:run -->
-      <plugin>
-        <groupId>org.apache.camel</groupId>
-        <artifactId>camel-maven-plugin</artifactId>
-        <version>2.18.0</version>
       </plugin>
     </plugins>
   </build>
