+++ services-camel/src/main/java/edu/unc/lib/cdr/AddDerivativeProcessor.java
@@ -1,0 +1,40 @@
+/**
+ * Copyright 2016 The University of North Carolina at Chapel Hill
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package edu.unc.lib.cdr;
+
+import org.apache.camel.Exchange;
+import org.apache.camel.Processor;
+
+import edu.unc.lib.dl.fcrepo4.Repository;
+
+/**
+ * Adds a derivative file to an existing file object
+ * 
+ * @author bbpennel
+ *
+ */
+public class AddDerivativeProcessor implements Processor {
+
+	private final Repository repository;
+	
+	public AddDerivativeProcessor(Repository repository) {
+		this.repository = repository;
+	}
+
+	@Override
+	public void process(Exchange exchange) throws Exception {
+	}
+}
