+++ services-camel/src/main/resources/log4j2.properties
@@ -8,5 +8,5 @@ appender.out.type = Console
 appender.out.name = out
 appender.out.layout.type = PatternLayout
 appender.out.layout.pattern = [%30.30t] %-30.30c{1} %-5p %m%n
-rootLogger.level = INFO
+rootLogger.level = TRACE
 rootLogger.appenderRef.out.ref = out
